+++ test/svlog/regr/iss199_modport.sv
@@ -1,5 +1,4 @@
 // RUN: moore %s -e top
-// IGNORE
 
 interface bus;
     logic clk;
@@ -23,3 +22,12 @@ module top;
     testcase test_inst0 (.the_bus(bus_inst0)); // this works
     testcase test_inst1 (.the_bus(bus_inst1.recv)); // this breaks
 endmodule
+
+// CHECK: entity @top () -> () {
+// CHECK:     %bus_inst0.clk = sig i1 %0
+// CHECK:     %bus_inst0.data = sig i1 %0
+// CHECK:     %bus_inst1.clk = sig i1 %0
+// CHECK:     %bus_inst1.data = sig i1 %0
+// CHECK:     inst @testcase.param3 (i1$ %bus_inst0.clk, i1$ %bus_inst0.data) -> (i1$ %test_inst0.result.default)
+// CHECK:     inst @testcase.param4 (i1$ %bus_inst1.clk, i1$ %bus_inst1.data) -> (i1$ %test_inst1.result.default)
+// CHECK: }